support for ython3

+++ volume.py
@@ -16,17 +16,6 @@ import subprocess
 from infi.dtypes.iqn import make_iscsi_name
 from time import gmtime, strftime
 
-#zone_file='./zones.json'
-#global zoneset
-#zoneset=get_zones_data(zone_file)
-#set_box_hexa(zoneset)
-#try:
-#    box_login(zoneset)
-#except Exception as E:
-    #logging.error("unable to login to infinibox, aborting {}".format(E))
-#    print E
-#    exit(5)
-
 ### Wrapper 
 def loggin_in_out(func):
     def wrapper(*args,**kwargs):
@@ -70,28 +59,13 @@ class InvalidUsage(Exception):
         rv['message'] = self.message
         return rv
 
-#app = Flask(__name__)
-#api = Api(app)
 
 
 ## To be replaced with the actual values
 loggedout_attempts=3
 loggedout_interval=3
-#ibox = "192.168.0.30"
-#notify_dir = '/tmp/'
-#notify_log = notify_dir+ "notify.log"
-#notify_script = "./notify_rm.sh"
-#cred=('admin', '123456')
 creds = HTTPBasicAuth('admin', '123456')
 
-### InfiniSDK Par/t
-#Creds=HTTPBasicAuth(cred)
-#ITAI 08112018
-#system=InfiniBox(ibox,cred)
-#system.login()
-#pool=system.pools.to_list()[0]
-#ITAI 08112018
-###
 # Constants
 onegig = 1000000000
 service_id='d4a44b0a-e3c2-4fec-9a3c-1d2cb14328f9'
@@ -104,11 +78,9 @@ vol_name_length=10
 
 
 ## Functions
-#generate_random_name=lambda length: ''.join(random.choice(string.ascii_letters + string.digits) for _ in range(length))
 ts=lambda now: strftime("%Y-%m-%d %H:%M:%S", gmtime())
 new_size = lambda  size: size/1000/1000/1000 
 def new_date(date):
-	#print "Date is {}".format(date)
 	tsa=arrow.get(str(date)[:-3])
 	return tsa.format(date_format)
 
@@ -116,13 +88,7 @@ def set_new_id(id):
 	embedding_zeros=id_len-len(str(id))
 	new_id=id_str+embedding_zeros*"0"+str(id)
 	return new_id
-#def notify_rm(file):
-#	try:
-#		return subprocess.Popen([notify_script,file])
-#	except Exception as E:
-#		notify_log_file=open(notify_log,'w')
-#		notify.write("Failed to call notify, {}".fomrat(E))
-#
+
 	
 def add_metadata(volume):
     ret_dict={}
@@ -132,13 +98,10 @@ def add_metadata(volume):
     return ret_dict
 
 
-
 def get_vol_data(volume):
     return_json={}
     return_json['volumes']={}
     return_json['volumes'].update(add_metadata(volume))
-    #return_json['volumes']['id'] = set_new_id(volume.get_id())
-    #return_json['volumes']['id']=encode_vol_by_id(val=volume)
     return_json['volumes']['size'] = volume.get_size().bits/8/1000000000
     return_json['volumes']['create_at'] = volume.get_created_at().format('YYYY-MM-DD HH:mm:ss')
     return_json['volumes']['lun_id'] = volume.get_id()
@@ -148,15 +111,14 @@ def get_vol_data(volume):
     else:
         return_json['volumes']['attach_status'] = 'offline'
     if 'status' in volume.get_all_metadata():
-	return_json['volumes']['status'] = volume.get_all_metadata()['status']
+        return_json['volumes']['status'] = volume.get_all_metadata()['status']
     else:
-	return_json['volumes']['status'] = 'available'
+        return_json['volumes']['status'] = 'available'
     return return_json
 
 class VolumesList(Resource):
     def __init__(self):
         self.reqparse = reqparse.RequestParser()
-        #self.reqparse.add_argument('pool_id', type=int, required=False, location='json',default=poolselect()['id'])
         self.reqparse.add_argument('name', type=str, required=True, location='json')
         self.reqparse.add_argument('provtype', type=str, required=False, location='json', default='THIN')
         self.reqparse.add_argument('size', type=int, required=True, location='json')
@@ -166,7 +128,6 @@ class VolumesList(Resource):
     def get(self):
         outp=[]
         return_json={}
-        ## ITAI 081118
         if 'iscsi_init' in request.args:
             iscsi_filter=request.args['iscsi_init']
         else:
@@ -174,101 +135,63 @@ class VolumesList(Resource):
         volumes=[]
         for box in zones['zones']:
             if ( isinstance(box['ibox'],InfiniBox)):
-                #print "box is {}".format(box['ibox'])
                 volumes.extend(box['ibox'].volumes.find(type='master').to_list())
-            #volumes=system.volumes.to_list()
             
-            for volume in volumes:
-                if iscsi_filter and 'iscsi_init' in volume.get_all_metadata().keys() and volume.get_metadata_value('iscsi_init') != iscsi_filter:
-                    continue 
-                else: 
-                   # print "vol is {}".format(volume.get_name())
-                    cur_vol=get_vol_data(volume)
-                    #print "cur vol is {} and its type is {}".format(cur_vol,type(cur_vol))
-                    #print "*******"
-                    #print len(outp)
-                    outp.append(cur_vol['volumes'])
-            ## ITAI 081118
-            #print outp
-            return_json['volumes']=outp
-            return return_json, 200
+        for volume in volumes:
+            if iscsi_filter and 'iscsi_init' in volume.get_all_metadata().keys() and volume.get_metadata_value('iscsi_init') != iscsi_filter:
+                continue 
+            else: 
+                cur_vol=get_vol_data(volume)
+                outp.append(cur_vol['volumes'])
+        return_json['volumes']=outp
+        return return_json, 200
     
     @loggin_in_out
     def post(self):
-
-        #body=request.json
         body=request.json
         for mandatory_key in mandatory_pars:
             if mandatory_key not in body['volumes']: ##1
-                print "mandatory_key {} can't be found".format(mandatory_key)
+                print("mandatory_key {} can't be found").format(mandatory_key)
                 raise InvalidUsage('Mandatory key cannot be found', status_code=410)
-        #try:
-            ## ITAI 081118
- 	#print "zoneset {} looking for {}".format(zoneset,body['volumes']['zone_code'])
         system=get_box_by_par(par="name",req="ibox",val=body['volumes']['zone_code'],zones=zones)
-        #print "******* found {}".format(system.get_name())
         pool=system.pools.to_list()[0]
-        #print "system is {}, pool is {}".format(system.get_name(),pool.get_name())
         if not system:
             return Response(status = 404)
-            ## ITAI 081118
         new_name=generate_random_name(vol_name_length)
         volume=system.volumes.create(pool=pool,size=body['volumes']['size']*GB,name=new_name)
-        #except Exception as E:
-        ##    print "exception is {}".format(E)
-        #    raise InvalidUsage('Error Caught {}'.format(E), status_code=420)
-	if body['volumes']['iscsi_init']:
-		host=get_host(system, body['volumes']['iscsi_init'])
-		host.map_volume(volume)
-		volume.set_metadata('status', 'in-use')	
+        if body['volumes']['iscsi_init']:
+            host=get_host(system, body['volumes']['iscsi_init'])
+            host.map_volume(volume)
+            volume.set_metadata('status', 'in-use')	
         volume.set_metadata('name',body['volumes']['name'])
         volume.set_metadata('iscsi_init',body['volumes']['iscsi_init'])
         volume.set_metadata('status','available')
         new_id=encode_vol_by_id(val=system,id=volume.get_id(),type='ibox',zones=zones)
-        #print ">>>> new id is {}".format(new_id)
         volume.set_metadata('id',new_id)
         for optional_key in opts_pars:
             if optional_key in body['volumes']:
                 volume.set_metadata(optional_key,body['volumes'][optional_key])
             else:
                 volume.set_metadata(optional_key,opts_pars[optional_key])
-        ## ITAI 081118
-        #vol_new_id=set_new_id(volume.get_id())
-        #vol_new_id=encode_vol_by_id(val=system.get_serial(),id=volume.get_id(),type='serial_dec')
-        #print "new id: {}".format(vol_new_id)
-        ## ITAI 081118
-        #notify=notify_dir+new_id
-        url="http://{}/api/rest/volumes/{}".format(system.get_name(),volume.get_id())
-        #print "url is {}".format(url)
-        vol_infi_data=requests.get(url=url,auth=creds)
+        # url="http://{}/api/rest/volumes/{}".format(system.get_name(),volume.get_id())
+        # vol_infi_data=requests.get(url=url,auth=creds)
         global iscsi_init
-        global volume_type
+        # global volume_type
         if 'iscsi_init' in body['volumes'].keys():
             iscsi_init=body['volumes']['iscsi_init']
         else:
             iscsi_init=''
-        #print "getting vol data"
         vol_data=get_vol_data(volume)
-        #print "vol data is {}".format(vol_data)
         notify_vol={}
-        #notify_vol={"snapshot_id":"", "notify_type":"volume_create","status":"available","result":"success"}
-	notify_vol={'volume_id':new_id, 'id':"", 'status':'available', 'notify_type':'volume_create'}
-        #notify_vol['volume_id'] = vol_new_id
-        #notify_vol['id']=new_id
-        #notify_vol['create_at'] = vol_data['volumes']['create_at']
+        notify_vol={'volume_id':new_id, 'id':"", 'status':'available', 'notify_type':'volume_create'}
         try:
-	    thread_a = NotifyRM(notify_vol)
-	    thread_a.start()            
-	   # notify_f=open(notify,'w')
-           # notify_f.writelines(json.dumps(notify_vol))
-           # notify_f.close()
-           # notify_rm(notify)
-	    #print "got until here"
+            thread_a = NotifyRM(notify_vol)
+            thread_a.start()            
         except Exception as E:
             str="Failed! {} ; notify is {}".format(E,notify)
-            print str 		
+            print(str) 		
             #return str,400
-	vol_data['volumes']['status']='creating'
+        vol_data['volumes']['status']='creating'
         return vol_data, 200
 
 class Volume(Resource):
@@ -278,18 +201,11 @@ class Volume(Resource):
 
     @loggin_in_out
     def get(self, vol_id):
-
-        #ITAI 08112018
-
-        ###infi_id=vol_id[-5:]
         system,vol=decode_vol_by_id(vol_id,'ibox',zones)
-        #ITAI 08112018
         try:
-            #print "looking on {} for {}".format(system.get_name(), vol)
             volume=system.volumes.find(id=vol)[0]
         except Exception: #outp_json['error'] or not outp_json['result']:
             return Response(status = 404)
-        #print "found vol {}".format(vol.get_name())
         return_json=get_vol_data(volume)
         return return_json, 200
         
@@ -298,48 +214,32 @@ class Volume(Resource):
 
     def put(self, id):
         body=request.json
-	#print type(body)
-	print body.keys()
+        print(body.keys())
         string="id is {} data is {}".format(id, body)
         return string,400
     
     @loggin_in_out
     def delete(self, vol_id):
-        #notify=notify_dir+vol_id
-        #print "in volume deletion"
-        #ITAI 08112018
-        ###infi_vol_id=int(vol_id[-5:])
         system,vol=decode_vol_by_id(vol_id,'ibox',zones)
-        #print "for deletion - volume {} in box {}".format(int(vol),system)
         try:
             volume=system.volumes.find(id=int(vol))
-            #print "found for deletion {}".format(volume.get_id())
             if volume:
-                vol_data=get_vol_data(volume[0])
-                volume[0].delete()
+                if volume[0].is_mapped():
+                    volume[0].unmap()
+                    volume[0].delete()
+                #vol_data=get_vol_data(volume[0])
+#                volume[0].delete()
             else: 
                 return Response(status = 200)
         except Exception as E:
-            print E
+            print(E)
             return E.message, 404
-        #ret_data={}
-        #ret_data['volume_id']=vol_id
-        #ret_data['create_at']=vol_data['volumes']['create_at']
-        #ret_data['status']='deleted'
-        #ret_data['result']='success'
-        #ret_data['snapshot_id']=""
-        #ret_data['notify_type']='volume_delete'
-	ret_data={'volume_id':vol_id, 'id':"", 'status':'deleted', 'notify_type':'volume_delete'}
+        ret_data={'volume_id':vol_id, 'id':"", 'status':'deleted', 'notify_type':'volume_delete'}
         try:
-		thread_b = NotifyRM(ret_data)
-		thread_b.start()	    	
-	#	notify_f=open(notify,'w')
-	#    	notify_f.writelines(json.dumps(ret_data))
-	#    	notify_f.close()
-	#    	notify_rm(notify)
-	except Exception:
-	    	pass
-        #time.sleep(5)
+            thread_b = NotifyRM(ret_data)
+            thread_b.start()	    	
+        except Exception:
+            pass
         return Response(status = 200)
 
 class VolumesAttachment(Resource):
@@ -350,39 +250,31 @@ class VolumesAttachment(Resource):
     def post(self):
         body=request.json
         status='success'
-	#ITAI 08112018
-        ###host=get_host(system,body['volume']['iscsi_init'])
         for volume in body['volume']['volumes']:
-            #ITAI 08112018
             system,vol_inf_id=decode_vol_by_id(volume['volume_id'],'ibox',zones)
             host=get_host(system,body['volume']['iscsi_init'])
-            #ITAI 08112018
             vol=system.volumes.find(id=vol_inf_id).to_list()
             if not vol:
                 pass
             if body['volume']['action'].upper() == "ATTACH":
-                try: 
-                    #host.map_volume(vol[0],lun=volume['order'])
-			
+                try: 			
                     host.map_volume(vol[0])
-		    #vol[0].set_metadata('status', 'in-use')
                 except Exception as E:
-                    print "Execption {}".format(E)
+                    print("Execption {}").format(E)
                     status='fail'
             elif body['volume']['action'].upper() == "DETACH":
                 ## TASK - Add tests here to find if volume is 'in use'
-                for attempt in xrange(loggedout_attempts):
+                for val in range(loggedout_attempts):
+                # for attempt in xrange(loggedout_attempts):
                     val=check_iqn_logged_in(system,body['volume']['iscsi_init'])
-                    #print "val is {}".format(val)
                     if val:
-                         #print "unmapping {} which is {}".format(vol[0],type(vol[0]))
-                         host.unmap_volume(vol[0])
-			 vol[0].set_metadata('status', 'available')
-                         status='success'
-                         break
+                        host.unmap_volume(vol[0])
+                        vol[0].set_metadata('status', 'available')
+                        status='success'
+                        break
                     else:
                         time.sleep(loggedout_interval)
-                        print "Host is still online"
+                        print("Host is still online")
                         status='fail'
             else:
                 status='fail'
@@ -397,49 +289,27 @@ class VolumeExpand(Resource):
     @loggin_in_out        
     def post(self,vol_id):
         body=request.json
-        #print "body is {}".format(body)
-        #print "id is {}".format(vol_id)
-        #volume=body['volume']['volume_id']
-        volume=vol_id
+        #volume=vol_id
         new_size=int(body['volume']['size'])
-        #ITAI 08112018
         system,vol_inf_id=decode_vol_by_id(vol_id,'ibox',zones)
         volume_object=system.volumes.find(id=vol_inf_id).to_list()
         if not volume_object:
             return 'Volume Not Found', 404
         volume_size=volume_object[0].get_size().bits/8/1000000000
-        #print "volume size is {} new size is {}".format(volume_size,new_size)
         if volume_size > new_size:
             return 'Volume is already bigger', 405
         cap_to_resize=(new_size-volume_size)*GB
         try:
             volume_object[0].resize(cap_to_resize)
         except Exception as E:
-            print "Caught Exception {}".format(E)
+            print("Caught Exception {}").format(E)
             return 'Exception', 500
-        #ret_data={}
-        #ret_data['volume_id']=vol_id
-        #ret_data['snapshot_id']=''
-        #ret_data['status']='available'
-        #ret_data['result']='success'
-        #ret_data['notify_type']='volume_extend'
-        #ret_data['create_at']=ts("now")
-	ret_data={'volume_id':vol_id, 'id':"", 'status':'available', 'notify_type':'volume_extend'}
+        ret_data={'volume_id':vol_id, 'id':"", 'status':'available', 'notify_type':'volume_extend'}
         try:
-            #print "notify is {}".format(notify)
-	    thread_a = NotifyRM(ret_data)
+            thread_a = NotifyRM(ret_data)
             thread_a.start()  
- #           notify_f=open(notify,'w')
- #           notify_f.writelines(json.dumps(ret_data))
- #           notify_f.close()
- #           notify_rm(notify)
         except Exception:
             pass
         return Response(status = 200)
     
-#api.add_resource(VolumesList, "/api/v1/volumes")
-#api.add_resource(VolumesAttachment, "/api/v1/volumes/attachment")
-#api.add_resource(Volume, "/api/v1/volumes/<string:vol_id>")
-#api.add_resource(VolumeExpand, "/api/v1/volumes/<string:vol_id>/expand")
-#app.run(debug=True, port=8080, host='0.0.0.0')
